Add a toggle for the HOME_ICON
Prior to this, the HOME_ICON was always visible when using an Awesome or
flat font. Some users might want to hide that icon to save a bit of
horizontal space. Also, as the HOME_ICON is visible even when not in a
home directory, it can be a bit misleading if a user were to think that
it meant they were currently in their home directory (essentially
duplicating the meaning of '~').

This commit adds a toggle called `POWERLEVEL9K_SHOW_HOME_ICON` that, when
`true`, will hide the icon from the `dir` prompt.