Add pyenv segment showing active python version
I created a new segment for use with [pyenv](https://github.com/yyuu/pyenv) to
manage python versions. This prompt can be used in place of `virtualenv`
segment, however, it is targeted at users of pyenv and not meant to
replace the `virtualenv` segment.

The prompt works by parsing output of `pyenv version` and displaying the
first word of the output as segment text. The design (color etc.) is
same as the `virtualenv` segment to respect consistency.

Tested on my personal machines (Mac OSX 10.11.4 and Ubuntu 15.04). The
segment would need to be revisited if `pyenv version` changes its output
format.