Improve joining of segments
Take the segment specific condition into account, when deciding if a
segment should be printed or not. This is better than just checking the
mere existance of last segments content, but still is not sufficient.
For example if we have segments A, B and C all of them are joined, but
B's condition is false, then C will be promoted what is not right.