Add option `PROJECT_ONLY` to node_version segment.
If this option is set, the segment is only shown inside node projects.
Such a project is identified by a package.json description.
The search is reverse from the current working diretory to the root.
The search stops if such a description has been found.
Add condition to show the segment, depending on the option variable and
the search result.