SEPARATION OF SEGMENTS INTO OWN FILES
NEW:
This PR moves all the segments into their own files.
It adds the ability to register icons dynamically, thus making icon
management the responsibility of the segment. This also allows us to
only load the icons we require instead of loading all the icons.

Utilities: union - returns the union of two arrays

UPDATES:
Split OS detection logic (in utilities) and recoded PROMPT_OS to make
use of the split.

TODO:
Update all the comments in all the files to ensure uniformity.

NOTES:
This is also the first step in preparing for a web configurator.