Improve upsearch
This improves upsearch in a way that it does not use recursion and uses
only a single `find` command to find all occurrences of the searched
item in the parent hierarchy.

The function now returns the results in opposite order, so that you get
the longest match first. This is what you probably want when truncating
a directory.
Also, the users home directory is only returned, if there was a match.
Previous:
/Users/dr
/Users/dr/test/1
/Users/dr/test/1/12/123

Now:
/Users/dr/test/1/12/123
/Users/dr/test/1