Remove check for numerical parameters in isSameColor
We now switched from strings ("none", "initial") to negative, numerical
values. This way, we don't have to check if given parameters are
numerical in isSameColor. This also means, that we use "-2" as initial
values for CURRENT_BG and CURRENT_RIGHT_BG.
As a side effect we had to change the check for numerical values in
getColorCode, as this did not work with negative values.