Fix segment loading
This fixes segment loading, even if:
- the segment is joined
- no prompt elements were defined by the user upfront

If the segment is a custom one, we skip the loading mechanism. Custom
segments must be loaded by the user.

Additionally, the logic was turned around: Before we iterated over all
possible segments and checking which one must be loaded. Now we just
iterate over all loaded segments.