Fix broken prompt when quickly pressing ctrl-c
The fix for double expansion of prompt was to delay the rendering
of the prompt as much as possible, and leave it to ZSH to escape
everything correctly. This led $PROMPT to just contain
'$(__p9k_build_left_prompt)'. Before it contained the contents of
that evaluation. The current fix is now to quote all input values
coming from VCS_INFO and work on the quoted values.