Fix error if git dir contains special characters
If the directory of the repo checkout contains umlauts or CJK
characters, the check for untracked files raised an error
"fatal: <directory> is outside repository". This was because
the directory name given to git rev-parse --show-toplevel was
not quoted enough. Interestingly, real quoting "(q)" did not
fix the issue, but adding single quotes did.