Enhacement: Go Version segment uses Go colours and only the version number is shown.
1. Why is this change neccesary?
Because the go_version segment had a green background when other segments (like
rbenv) have a colour that matches the language.
The Go version had the string "go" before the semantic versioning being
redundant to have the Gopher icon and the String.
The segment was only visible when inside the GOPATH, and some users (like me)
may be interested in the segment being shown persistently.

2. How does it address the issue?
The background colour was changed from green to cyan and the foreground colour
from white to black in order to match Go colours.
Only the version of Go is shown in the segment.
A variable was added to see the segment persistently if the user wishes to.

3. What side effects does this change have?
None!