More consistent results from 'ip' on Non-OSX systems
Due to issues matching with some versions of iproute's 'ip' command, I've rewritten and thoroughly commented. See https://github.com/Powerlevel9k/powerlevel9k/issues/823

If desired, a lot of this can be rewritten to be shorter (and theoretically a smidge faster to execute); I chose to write the initial code in this manner to make it easier to read what I've done when judging the pull request.